vkd3d: Replace atomic_add_fetch() by InterlockedAdd().
By analogy with InterlockedIncrement. It avoids the need for a
configure check on Windows platforms.

+++ libs/vkd3d/vkd3d_private.h
@@ -192,7 +192,7 @@ struct vkd3d_fence_worker
     bool should_exit;
     bool pending_fence_destruction;
 
-    size_t enqueued_fence_count;
+    LONG enqueued_fence_count;
     struct vkd3d_enqueued_fence
     {
         VkFence vk_fence;
